TimeToTravel + AgeCalculator Added to Ship Class

+++ Team-Project-Week-4/Player.cs
@@ -13,7 +13,7 @@ using Team_Project_Week_4;
 
 public class Player
 {
-    public int playerAge;
+    public double playerAge;
     public (int playerX, int playerY) playerLocation;
     public int playerX;
     public int playerY;
